restructure and enhance RadioButtonGroup validation tests
