+++ vaadin-radio-button-flow-parent/vaadin-radio-button-flow-integration-tests/src/test/java/com/vaadin/flow/component/radiobutton/tests/RadioButtonGroupIT.java
@@ -260,22 +260,4 @@ public class RadioButtonGroupIT extends AbstractComponentIT {
         $(TestBenchElement.class).id("clear-helper-component-button").click();
         Assert.assertNull(groupWithHelperComponent.getHelperComponent());
     }
-
-    private void verifyGroupInvalid(TestBenchElement group,
-            TestBenchElement errorMessage) {
-        Assert.assertEquals("Radio button group is invalid.", true,
-                group.getPropertyBoolean("invalid"));
-        Assert.assertEquals("Error message should be shown.",
-                Boolean.FALSE.toString(),
-                errorMessage.getAttribute("aria-hidden"));
-    }
-
-    private void verifyGroupValid(TestBenchElement group,
-            TestBenchElement errorMessage) {
-        Assert.assertEquals("Radio button group is not invalid.", false,
-                group.getPropertyBoolean("invalid"));
-        Assert.assertEquals("Error message should be hidden.",
-                Boolean.TRUE.toString(),
-                errorMessage.getAttribute("aria-hidden"));
-    }
 }
